requirements.txt
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
fiona==1.9.6
Flask==3.0.2
geographiclib==2.0
geopandas==0.14.3
geopy==2.4.1
google-auth==2.29.0
google-auth-oauthlib==1.2.0
gspread==6.1.0
httplib2==0.22.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
packaging==24.0
pandas==2.2.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
pyproj==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
shapely==2.0.3
six==1.16.0
StrEnum==0.4.15
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.2
